# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Care Platform API",
        "description": "Complete API documentation for Care Platform - A comprehensive marketplace connecting service providers with customers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Care Platform API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/payouts": {
            "get": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Get all payout requests",
                "operationId": "b5ef720b412c8b96870cb42bff2d1539",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/payouts/{id}": {
            "get": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Get payout details",
                "operationId": "22ef6e48da3e0d6d5c6cb5fb8478faa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/payouts/{id}/approve": {
            "post": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Approve payout request",
                "operationId": "b93c3e423a7353b0ad1306b43f0d9557",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approved successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/payouts/{id}/reject": {
            "post": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Reject payout request",
                "operationId": "c328957022723c2ab24460863e4dff17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rejected successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/payouts/statistics": {
            "get": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Get payout statistics",
                "operationId": "3bb91b7e4a3ec1aac0e42c2703ce10d5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/payouts/bulk-approve": {
            "post": {
                "tags": [
                    "Admin - Payouts"
                ],
                "summary": "Bulk approve payouts",
                "operationId": "ccf1756e4edef7e204591ca69f5c7cec",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}